chore: speed up the labware stackup tests somewhat
The labware stackup integration tests rule but were also very slow,
because they were running serially and there are 8,000 of them. By
making them run in parallel using
concurrent.futures.ProcessPoolExecutor, we can make it take like a
minute instead of 2 hours.
