+++ api/.gitignore
@@ -4,3 +4,4 @@ build
 htmlcov
 
 src/opentrons/_version.py
+tests/**/*.temp
